# shared/api/order/v1/order.openapi.yaml
openapi: 3.1.0
info:
  title: Order Service API
  version: 1.0.0
  description: API для управления заказами на постройку космических кораблей.

servers:
  - url: /api/v1

paths:
  /orders:
    post:
      summary: Создать заказ
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '200':
          description: Заказ успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateOrderResponse'
        '400':
          description: Некорректный запрос
        '404':
          description: Одна или несколько деталей не найдены

  /orders/{order_uuid}/pay:
    post:
      summary: Оплатить заказ
      operationId: payOrder
      parameters:
        - name: order_uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PayOrderRequest'
      responses:
        '200':
          description: Заказ успешно оплачен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PayOrderResponse'
        '404':
          description: Заказ не найден

  /orders/{order_uuid}:
    get:
      summary: Получить заказ по UUID
      operationId: getOrder
      parameters:
        - name: order_uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Информация о заказе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Заказ не найден

  /orders/{order_uuid}/cancel:
    post:
      summary: Отменить заказ
      operationId: cancelOrder
      parameters:
        - name: order_uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Заказ успешно отменён
        '404':
          description: Заказ не найден
        '409':
          description: Заказ уже оплачен и не может быть отменён

components:
  schemas:
    CreateOrderRequest:
      type: object
      required: [user_uuid, part_uuids]
      properties:
        user_uuid:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        part_uuids:
          type: array
          maxItems: 20
          items:
            type: string
            format: uuid
          example:
            - "111e2222-e89b-12d3-a456-426614174001"
            - "222e3333-e89b-12d3-a456-426614174002"

    CreateOrderResponse:
      type: object
      required: [order_uuid, total_price]
      properties:
        order_uuid:
          type: string
          format: uuid
          example: "444e5555-e89b-12d3-a456-426614174004"
        total_price:
          type: number
          format: double
          example: 123.45

    PayOrderRequest:
      type: object
      required: [payment_method]
      properties:
        payment_method:
          $ref: '#/components/schemas/PaymentMethod'

    PayOrderResponse:
      type: object
      required: [transaction_uuid]
      properties:
        transaction_uuid:
          type: string
          format: uuid
          example: "666e7777-e89b-12d3-a456-426614174006"

    Order:
      type: object
      required: [order_uuid, user_uuid, part_uuids, total_price, status]
      properties:
        order_uuid:
          type: string
          format: uuid
          example: "444e5555-e89b-12d3-a456-426614174004"
        user_uuid:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        part_uuids:
          type: array
          maxItems: 20
          items:
            type: string
            format: uuid
          example:
            - "111e2222-e89b-12d3-a456-426614174001"
        total_price:
          type: number
          format: double
          example: 123.45
        transaction_uuid:
          type: string
          format: uuid
          nullable: true
          example: "666e7777-e89b-12d3-a456-426614174006"
        payment_method:
          $ref: '#/components/schemas/PaymentMethod'
        status:
          $ref: '#/components/schemas/OrderStatus'

    PaymentMethod:
      type: string
      enum:
        - UNKNOWN
        - CARD
        - SBP
        - CREDIT_CARD
        - INVESTOR_MONEY
      example: CARD

    OrderStatus:
      type: string
      enum:
        - PENDING_PAYMENT
        - PAID
        - CANCELLED
      example: PENDING_PAYMENT